requirements.txt
psycopg2==2.8.6
psycopg2-binary==2.8.6
django==3.1.1
django-model-utils==4.0.0
requests==2.24.0
flake8==3.8.4
backoff==1.10.0
python-dotenv==0.14.0
gunicorn==20.0.4